fixed up PR suggestions
