+++ src/main/java/uk/gov/hmcts/reform/em/hrs/util/CvpConnectionResolver.java
@@ -2,6 +2,9 @@ package uk.gov.hmcts.reform.em.hrs.util;
 
 import lombok.NoArgsConstructor;
 
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 import static lombok.AccessLevel.PRIVATE;
 
 @NoArgsConstructor(access = PRIVATE)
@@ -12,5 +15,25 @@ public class CvpConnectionResolver {
         boolean isACvpEndpointUrl =
             cvpConnectionString.contains("cvprecordings") && !cvpConnectionString.contains("AccountName");
         return isACvpEndpointUrl;
+    }
+
+
+    /*
+       url pattern is expected to be either
+    https://cvprecordingsstgsa-secondary.blob.core.windows.net/
+       or
+    https://cvprecordingsstgsa.blob.core.windows.net/
+    */
+    static Pattern pattern = Pattern.compile("https://(.*?)(?:-secondary)?.blob.core.windows.net");
+
+    static String extractAccountFromUrl(String cvpConnectionString) {
+
+        Matcher matcher = pattern.matcher(cvpConnectionString);
+        String accountName = null;
+
+        if (matcher.find()) {
+            accountName = matcher.group(1);
+        }
+        return accountName;
     }
 }
