+++ src/test/java/uk/gov/hmcts/reform/em/hrs/util/CvpConnectionResolverTest.java
@@ -2,15 +2,16 @@ package uk.gov.hmcts.reform.em.hrs.util;
 
 import org.junit.jupiter.api.Test;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static uk.gov.hmcts.reform.em.hrs.util.CvpConnectionResolver.extractAccountFromUrl;
 import static uk.gov.hmcts.reform.em.hrs.util.CvpConnectionResolver.isACvpEndpointUrl;
 
 public class CvpConnectionResolverTest {
     @Test
     public void testIsACvpEndpointUrl() {
-        //SECURITY NOTICE The below connection string / account key contains no sensitive information, and is derived
-        //from the publicly known developer key - https://docs.microsoft.com/en-us/azure/storage/common/storage-configure-connection-string
+
         assertFalse(isACvpEndpointUrl(
             "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;"
                 + "AccountKey=ACTUALKEYNOTNEEDEDFORTHISTEST"
@@ -24,6 +25,23 @@ public class CvpConnectionResolverTest {
         assertTrue(isACvpEndpointUrl("https://cvprecordingssboxsa.blob.core.windows.net/"));
         assertTrue(isACvpEndpointUrl("https://cvprecordingsstgsa.blob.core.windows.net/"));
         assertTrue(isACvpEndpointUrl("https://cvprecordingsprodsa.blob.core.windows.net/"));
+    }
+
+    @Test
+    void testUrlWithSecondaryDNSWorks() {
+        String input = "https://cvprecordingsstgsa-secondary.blob.core.windows.net/";
+        String expected = "cvprecordingsstgsa";
+        String actual = extractAccountFromUrl(input);
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    void testUrlWithPrimaryDNSWorks() {
+        String input = "https://cvprecordingsstgsa.blob.core.windows.net/";
+        String expected = "cvprecordingsstgsa";
+        String actual = extractAccountFromUrl(input);
+        assertEquals(expected, actual);
+
     }
 }
